Cleaned up files, fixed a couple insignificant warnings

+++ ros2/speech_ros2/src/string_to_gcode.hpp
@@ -1621,7 +1621,7 @@ std::vector<std::vector<xy>> gcodeVector = {{},
                                             {0, 0, "G01"},
                                             {50, 0, "G01"}},
 
-                                            //\
+                                            //backslash
 
                                             {{0, 50, "G00"},
                                             {15.55555, 0, "G01"}},
@@ -2397,7 +2397,7 @@ std::vector<std::vector<xy>> gcodeVector = {{},
     if(placeY > 25000 - (size * 50)){
         placeY = 25000 - (size * 50);
     }
-    for(unsigned int i = 0; i < input.size(); i++){
+    for(int i = 0; i < (int)input.size(); i++){
         if(!enters.empty()){
             if(i == enters[0]){
                 placeY -= 70 * size;
